refactor: rename 2 components to PascalCase
