+++ app/components/ExplorerMetrics.tsx
